refactor: 리프레시 토큰 저장소 명명 정리
