+++ src/main/java/com/techfork/auth/security/handler/login/OAuth2LoginRefreshTokenWriter.java
@@ -1,6 +1,6 @@
 package com.techfork.auth.security.handler.login;
 
-import com.techfork.auth.security.service.RefreshTokenService;
+import com.techfork.auth.security.token.RefreshTokenStore;
 import com.techfork.auth.security.util.CookieUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -11,13 +11,13 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 class OAuth2LoginRefreshTokenWriter {
 
-    private final RefreshTokenService refreshTokenService;
+    private final RefreshTokenStore refreshTokenStore;
 
     @Value("${server.domain}")
     private String domain;
 
     public void write(Long userId, OAuth2LoginTokens tokens, HttpServletResponse response) {
-        refreshTokenService.saveRefreshToken(userId, tokens.refreshToken(), tokens.refreshTokenExpiration());
+        refreshTokenStore.saveRefreshToken(userId, tokens.refreshToken(), tokens.refreshTokenExpiration());
         CookieUtil.addRefreshTokenCookie(response, domain, tokens.refreshToken(), tokens.refreshTokenExpiration());
     }
 }
