+++ src/main/java/com/techfork/auth/security/token/RefreshTokenStore.java
@@ -1,17 +1,16 @@
-package com.techfork.auth.security.service;
+package com.techfork.auth.security.token;
 
 import com.techfork.global.constant.RedisKey;
+import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.stereotype.Service;
-
-import java.util.concurrent.TimeUnit;
+import org.springframework.stereotype.Component;
 
 @Slf4j
-@Service
+@Component
 @RequiredArgsConstructor
-public class RefreshTokenService {
+public class RefreshTokenStore {
 
     private final StringRedisTemplate redisTemplate;
